[WNMGDS-966] badge typescript conversion (#1083)
* convert Badge to typescript

* removing commented out import in alert

* linting updates
