+++ packages/design-system-docs/src/pages/components/Badge/_Badge.docs.scss
@@ -15,7 +15,7 @@ Style guide: components.badge
 
 @react-example Badge.example.jsx
 
-@react-props Badge.jsx
+@react-props Badge.tsx
 
 Style guide: components.badge.react
 */
