+++ packages/design-system/src/types/index.d.ts
@@ -3,7 +3,6 @@
  */
 
 export { default as Autocomplete } from './Autocomplete/Autocomplete';
-export { default as Badge } from './Badge/Badge';
 export { default as Button } from './Button/Button';
 export { default as Dialog } from './Dialog/Dialog';
 export { default as FormLabel } from './FormLabel/FormLabel';
@@ -30,6 +29,7 @@ export { default as VerticalNav } from './VerticalNav/VerticalNav';
 // Type definitions generated by cmsds scripts
 export * from './Accordion';
 export * from './Alert';
+export * from './Badge';
 export * from './ChoiceList';
 export * from './DateField';
 export * from './Dropdown';
